ORCID: get all publications (using rorcid::works())

ORCID: get all publications (using rorcid::works())

create_publications_df_for_orcids(orcids = get_kwb_orcids())

Arguments

orcids
character vector with valid ORCIDs (default: get_kwb_orcids())

Value

data.frame with all (public!) ORCID publications for provided ORCIDS

Examples

orcids <- get_kwb_orcids()
create_publications_df_for_orcids(orcids)

#>                    orcid given_names family_name
#>   1: 0000-0001-5483-4594     Andreas   Matzinger
#>   2: 0000-0001-5483-4594     Andreas   Matzinger
#>   3: 0000-0001-5483-4594     Andreas   Matzinger
#>   4: 0000-0001-5483-4594     Andreas   Matzinger
#>   5: 0000-0001-5483-4594     Andreas   Matzinger
#>  ---                                            
#> 114: 0000-0002-5252-4880     Nicolas     Caradot
#> 115: 0000-0002-5252-4880     Nicolas     Caradot
#> 116: 0000-0002-5252-4880     Nicolas     Caradot
#> 117: 0000-0002-5252-4880     Nicolas     Caradot
#> 118: 0000-0002-7436-8575    Wolfgang        Seis
#>                                            url.x
#>   1:                                            
#>   2:                                            
#>   3:                                            
#>   4:                                            
#>   5:                                            
#>  ---                                            
#> 114: https://www.linkedin.com/in/nicolascaradot/
#> 115: https://www.linkedin.com/in/nicolascaradot/
#> 116: https://www.linkedin.com/in/nicolascaradot/
#> 117: https://www.linkedin.com/in/nicolascaradot/
#> 118:                                            
#>                                                        orcid_link put-code
#>   1: [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 88432325
#>   2: [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 88432324
#>   3: [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 88432326
#>   4: [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 88432323
#>   5: [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295084
#>  ---                                                                      
#> 114: [0000-0002-5252-4880](https://orcid.org/0000-0002-5252-4880) 86457954
#> 115: [0000-0002-5252-4880](https://orcid.org/0000-0002-5252-4880) 86457972
#> 116: [0000-0002-5252-4880](https://orcid.org/0000-0002-5252-4880) 86457963
#> 117: [0000-0002-5252-4880](https://orcid.org/0000-0002-5252-4880) 86457974
#> 118: [0000-0002-7436-8575](https://orcid.org/0000-0002-7436-8575)       NA
#>                  type visibility                               path
#>   1:  journal-article     public /0000-0001-5483-4594/work/88432325
#>   2:  journal-article     public /0000-0001-5483-4594/work/88432324
#>   3:  journal-article     public /0000-0001-5483-4594/work/88432326
#>   4:             book     public /0000-0001-5483-4594/work/88432323
#>   5:  journal-article     public /0000-0001-5483-4594/work/46295084
#>  ---                                                               
#> 114: conference-paper     public /0000-0002-5252-4880/work/86457954
#> 115:  journal-article     public /0000-0002-5252-4880/work/86457972
#> 116:  journal-article     public /0000-0002-5252-4880/work/86457963
#> 117:  journal-article     public /0000-0002-5252-4880/work/86457974
#> 118:             <NA>       <NA>                               <NA>
#>      display-index created-date.value last-modified-date.value
#>   1:             0       1.612884e+12             1.612884e+12
#>   2:             0       1.612884e+12             1.612884e+12
#>   3:             0       1.612884e+12             1.612884e+12
#>   4:             0       1.612884e+12             1.612884e+12
#>   5:             0       1.530808e+12             1.612884e+12
#>  ---                                                          
#> 114:             0       1.610116e+12             1.610116e+12
#> 115:             0       1.610116e+12             1.610116e+12
#> 116:             0       1.610116e+12             1.610116e+12
#> 117:             0       1.610116e+12             1.610116e+12
#> 118:          <NA>                 NA                       NA
#>      source.source-orcid source.assertion-origin-client-id
#>   1:                  NA                                NA
#>   2:                  NA                                NA
#>   3:                  NA                                NA
#>   4:                  NA                                NA
#>   5:                  NA                                NA
#>  ---                                                      
#> 114:                  NA                                NA
#> 115:                  NA                                NA
#> 116:                  NA                                NA
#> 117:                  NA                                NA
#> 118:                  NA                                NA
#>                       source.source-client-id.uri source.source-client-id.path
#>   1: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#>   2: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#>   3: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#>   4: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#>   5: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#>  ---                                                                          
#> 114: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 115: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 116: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 117: https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 118:                                         <NA>                         <NA>
#>      source.source-client-id.host source.source-name.value
#>   1:                    orcid.org        Scopus - Elsevier
#>   2:                    orcid.org        Scopus - Elsevier
#>   3:                    orcid.org        Scopus - Elsevier
#>   4:                    orcid.org        Scopus - Elsevier
#>   5:                    orcid.org        Scopus - Elsevier
#>  ---                                                      
#> 114:                    orcid.org        Scopus - Elsevier
#> 115:                    orcid.org        Scopus - Elsevier
#> 116:                    orcid.org        Scopus - Elsevier
#> 117:                    orcid.org        Scopus - Elsevier
#> 118:                         <NA>                     <NA>
#>          source.assertion-origin-orcid.uri source.assertion-origin-orcid.path
#>   1: https://orcid.org/0000-0001-5483-4594                0000-0001-5483-4594
#>   2: https://orcid.org/0000-0001-5483-4594                0000-0001-5483-4594
#>   3: https://orcid.org/0000-0001-5483-4594                0000-0001-5483-4594
#>   4: https://orcid.org/0000-0001-5483-4594                0000-0001-5483-4594
#>   5: https://orcid.org/0000-0001-5483-4594                0000-0001-5483-4594
#>  ---                                                                         
#> 114: https://orcid.org/0000-0002-5252-4880                0000-0002-5252-4880
#> 115: https://orcid.org/0000-0002-5252-4880                0000-0002-5252-4880
#> 116: https://orcid.org/0000-0002-5252-4880                0000-0002-5252-4880
#> 117: https://orcid.org/0000-0002-5252-4880                0000-0002-5252-4880
#> 118:                                  <NA>                               <NA>
#>      source.assertion-origin-orcid.host source.assertion-origin-name.value
#>   1:                          orcid.org                  Andreas Matzinger
#>   2:                          orcid.org                  Andreas Matzinger
#>   3:                          orcid.org                  Andreas Matzinger
#>   4:                          orcid.org                  Andreas Matzinger
#>   5:                          orcid.org                  Andreas Matzinger
#>  ---                                                                      
#> 114:                          orcid.org                    Nicolas Caradot
#> 115:                          orcid.org                    Nicolas Caradot
#> 116:                          orcid.org                    Nicolas Caradot
#> 117:                          orcid.org                    Nicolas Caradot
#> 118:                               <NA>                               <NA>
#>      title.subtitle title.translated-title
#>   1:             NA                     NA
#>   2:             NA                     NA
#>   3:             NA                     NA
#>   4:             NA                     NA
#>   5:             NA                     NA
#>  ---                                      
#> 114:             NA                     NA
#> 115:             NA                     NA
#> 116:             NA                     NA
#> 117:             NA                     NA
#> 118:             NA                     NA
#>                                                                                                                                                                                                                                                                         title.title.value
#>   1:                                                                                                                                                                  Nutrient retention by the littoral vegetation of a large lake: Can Lake Ohrid cope with current and future loading?
#>   2:                                                                                                                                                                Spatial compatibility of implementing nature-based solutions for reducing urban heat islands and stormwater pollution
#>   3:                                                                                                                                           Sustainable urban drainage systems in established city developments: Modelling the potential for CSO reduction and river impact mitigation
#>   4:                                                                                                                                                                                                       Relevance of Different CSO Outlets for Bathing Water Quality in a River System
#>   5:                                                                                                                                                             Impacts of combined sewer overflows on a large urban river – Understanding the effect of different management strategies
#>  ---                                                                                                                                                                                                                                                                                     
#> 114:                                                                                                                                                                                      Validation and sensitivity of a coupled model tool for CSO impact assessment in Berlin, Germany
#> 115:                                                                                                                                                                                                                           Urban flood risk assessment using sewer flooding databases
#> 116: Risk assessment of sewerage systems overflows: Methodological contributions from two case studies (Lyon and Mulhouse, France),L'évaluation du risque de débordement des systèmes de gestion des eaux urbaines: Contributions méthodologiques de deux cas d'études (Lyon et Mulhouse)
#> 117:                                               Sustainable management of urban water. Presentation of an adaptative model based on local assessment of service quality,Gestion durable des eaux urbaines et qualité de service Présentation d'un modèle adaptatif d'évaluation locale
#> 118:                                                                                                                                                                                                                                                                                 <NA>
#>      external-ids.external-id
#>   1:        <data.frame[3x7]>
#>   2:        <data.frame[3x7]>
#>   3:        <data.frame[3x7]>
#>   4:        <data.frame[3x7]>
#>   5:        <data.frame[2x7]>
#>  ---                         
#> 114:        <data.frame[1x7]>
#> 115:        <data.frame[3x9]>
#> 116:        <data.frame[3x9]>
#> 117:        <data.frame[3x9]>
#> 118:                         
#>                                                                              url.value
#>   1: http://www.scopus.com/inward/record.url?eid=2-s2.0-85085612085&partnerID=MN8TOARS
#>   2: http://www.scopus.com/inward/record.url?eid=2-s2.0-85090223918&partnerID=MN8TOARS
#>   3: http://www.scopus.com/inward/record.url?eid=2-s2.0-85089484020&partnerID=MN8TOARS
#>   4: http://www.scopus.com/inward/record.url?eid=2-s2.0-85071536629&partnerID=MN8TOARS
#>   5: http://www.scopus.com/inward/record.url?eid=2-s2.0-84986616713&partnerID=MN8TOARS
#>  ---                                                                                  
#> 114: http://www.scopus.com/inward/record.url?eid=2-s2.0-84894194693&partnerID=MN8TOARS
#> 115: http://www.scopus.com/inward/record.url?eid=2-s2.0-80052069733&partnerID=MN8TOARS
#> 116: http://www.scopus.com/inward/record.url?eid=2-s2.0-77954316987&partnerID=MN8TOARS
#> 117: http://www.scopus.com/inward/record.url?eid=2-s2.0-70249148000&partnerID=MN8TOARS
#> 118:                                                                              <NA>
#>      publication-date.month publication-date.day publication-date.year.value
#>   1:                     NA                   NA                        2020
#>   2:                     NA                   NA                        2020
#>   3:                     NA                   NA                        2020
#>   4:                     NA                   NA                        2019
#>   5:                     NA                   NA                        2016
#>  ---                                                                        
#> 114:                     NA                   NA                        2012
#> 115:                     NA                   NA                        2011
#> 116:                     NA                   NA                        2010
#> 117:                     NA                   NA                        2009
#> 118:                     NA                   NA                        <NA>
#>                                                                                                                                                 journal-title.value
#>   1:                                                                                                                                     Limnology and Oceanography
#>   2:                                                                                                                                   Sustainability (Switzerland)
#>   3:                                                                                                                            Journal of Environmental Management
#>   4:                                                                                                                                    Green Energy and Technology
#>   5:                                                                                                                                                 Water Research
#>  ---                                                                                                                                                               
#> 114: iEMSs 2012 - Managing Resources of a Limited Planet: Proceedings of the 6th Biennial Meeting of the International Environmental Modelling and Software Society
#> 115:                                                                                                                                   Water Science and Technology
#> 116:                                                                                                                               Techniques - Sciences - Methodes
#> 117:                                                                                                                               Techniques - Sciences - Methodes
#> 118:                                                                                                                                                           <NA>
#>      source.source-orcid.uri source.source-orcid.path source.source-orcid.host
#>   1:                    <NA>                     <NA>                     <NA>
#>   2:                    <NA>                     <NA>                     <NA>
#>   3:                    <NA>                     <NA>                     <NA>
#>   4:                    <NA>                     <NA>                     <NA>
#>   5:                    <NA>                     <NA>                     <NA>
#>  ---                                                                          
#> 114:                    <NA>                     <NA>                     <NA>
#> 115:                    <NA>                     <NA>                     <NA>
#> 116:                    <NA>                     <NA>                     <NA>
#> 117:                    <NA>                     <NA>                     <NA>
#> 118:                    <NA>                     <NA>                     <NA>
#>      publication-date.month.value publication-date.day.value
#>   1:                         <NA>                       <NA>
#>   2:                         <NA>                       <NA>
#>   3:                         <NA>                       <NA>
#>   4:                         <NA>                       <NA>
#>   5:                         <NA>                       <NA>
#>  ---                                                        
#> 114:                         <NA>                       <NA>
#> 115:                         <NA>                       <NA>
#> 116:                         <NA>                       <NA>
#> 117:                         <NA>                       <NA>
#> 118:                         <NA>                       <NA>
#>      source.source-client-id source.assertion-origin-orcid
#>   1:                      NA                            NA
#>   2:                      NA                            NA
#>   3:                      NA                            NA
#>   4:                      NA                            NA
#>   5:                      NA                            NA
#>  ---                                                      
#> 114:                      NA                            NA
#> 115:                      NA                            NA
#> 116:                      NA                            NA
#> 117:                      NA                            NA
#> 118:                      NA                            NA
#>      source.assertion-origin-name url.y journal-title
#>   1:                           NA    NA            NA
#>   2:                           NA    NA            NA
#>   3:                           NA    NA            NA
#>   4:                           NA    NA            NA
#>   5:                           NA    NA            NA
#>  ---                                                 
#> 114:                           NA    NA            NA
#> 115:                           NA    NA            NA
#> 116:                           NA    NA            NA
#> 117:                           NA    NA            NA
#> 118:                           NA    NA            NA