ORCID: get all publications (using rorcid::works())

ORCID: get all publications (using rorcid::works())

create_publications_df_for_orcids(orcids = get_kwb_orcids())

Arguments

orcids
character vector with valid ORCIDs (default: get_kwb_orcids())

Value

data.frame with all (public!) ORCID publications for provided ORCIDS

Examples

orcids <- get_kwb_orcids()
create_publications_df_for_orcids(orcids)

#>                  orcid given_names family_name
#> 1  0000-0001-5483-4594     Andreas   Matzinger
#> 2  0000-0001-5483-4594     Andreas   Matzinger
#> 3  0000-0001-5483-4594     Andreas   Matzinger
#> 4  0000-0001-5483-4594     Andreas   Matzinger
#> 5  0000-0001-5483-4594     Andreas   Matzinger
#> 6  0000-0001-5483-4594     Andreas   Matzinger
#> 7  0000-0001-5483-4594     Andreas   Matzinger
#> 8  0000-0001-5483-4594     Andreas   Matzinger
#> 9  0000-0001-5483-4594     Andreas   Matzinger
#> 10 0000-0001-5483-4594     Andreas   Matzinger
#> 11 0000-0001-5483-4594     Andreas   Matzinger
#> 12 0000-0001-5483-4594     Andreas   Matzinger
#> 13 0000-0001-5483-4594     Andreas   Matzinger
#> 14 0000-0001-5483-4594     Andreas   Matzinger
#> 15 0000-0001-5483-4594     Andreas   Matzinger
#> 16 0000-0001-5483-4594     Andreas   Matzinger
#> 17 0000-0001-5483-4594     Andreas   Matzinger
#> 18 0000-0001-5483-4594     Andreas   Matzinger
#> 19 0000-0001-5483-4594     Andreas   Matzinger
#> 20 0000-0001-5483-4594     Andreas   Matzinger
#> 21 0000-0001-5483-4594     Andreas   Matzinger
#> 22 0000-0001-5483-4594     Andreas   Matzinger
#> 23 0000-0001-5483-4594     Andreas   Matzinger
#> 24 0000-0001-5483-4594     Andreas   Matzinger
#> 25 0000-0001-5483-4594     Andreas   Matzinger
#> 26 0000-0001-9134-2871       Hauke  Sonnenberg
#> 27 0000-0001-9134-2871       Hauke  Sonnenberg
#> 28 0000-0001-9134-2871       Hauke  Sonnenberg
#> 29 0000-0001-9134-2871       Hauke  Sonnenberg
#> 30 0000-0001-9134-2871       Hauke  Sonnenberg
#> 31 0000-0001-9134-2871       Hauke  Sonnenberg
#> 32 0000-0001-9134-2871       Hauke  Sonnenberg
#> 33 0000-0002-7436-8575    Wolfgang        Seis
#> 34 0000-0003-0647-7726     Michael     Rustler
#> 35 0000-0003-0647-7726     Michael     Rustler
#> 36 0000-0003-0647-7726     Michael     Rustler
#> 37 0000-0003-0647-7726     Michael     Rustler
#> 38 0000-0003-0647-7726     Michael     Rustler
#> 39 0000-0003-0647-7726     Michael     Rustler
#> 40 0000-0003-0647-7726     Michael     Rustler
#> 41 0000-0003-0647-7726     Michael     Rustler
#> 42 0000-0003-0647-7726     Michael     Rustler
#> 43 0000-0003-0647-7726     Michael     Rustler
#>                                                                          url
#> 1                                                                           
#> 2                                                                           
#> 3                                                                           
#> 4                                                                           
#> 5                                                                           
#> 6                                                                           
#> 7                                                                           
#> 8                                                                           
#> 9                                                                           
#> 10                                                                          
#> 11                                                                          
#> 12                                                                          
#> 13                                                                          
#> 14                                                                          
#> 15                                                                          
#> 16                                                                          
#> 17                                                                          
#> 18                                                                          
#> 19                                                                          
#> 20                                                                          
#> 21                                                                          
#> 22                                                                          
#> 23                                                                          
#> 24                                                                          
#> 25                                                                          
#> 26                                                 https://github.com/hsonne
#> 27                                                 https://github.com/hsonne
#> 28                                                 https://github.com/hsonne
#> 29                                                 https://github.com/hsonne
#> 30                                                 https://github.com/hsonne
#> 31                                                 https://github.com/hsonne
#> 32                                                 https://github.com/hsonne
#> 33                                                                          
#> 34 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 35 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 36 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 37 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 38 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 39 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 40 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 41 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 42 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#> 43 http://www.mrustl.de , https://www.mendeley.com/profiles/michael-rustler/
#>                                                      orcid_link put-code
#> 1  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295084
#> 2  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295081
#> 3  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295078
#> 4  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295079
#> 5  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594)  5121191
#> 6  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295089
#> 7  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295080
#> 8  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295077
#> 9  [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295085
#> 10 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295087
#> 11 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594)  5121188
#> 12 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295086
#> 13 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295083
#> 14 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594)  5121192
#> 15 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594)  5121195
#> 16 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295090
#> 17 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594)  5121189
#> 18 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295088
#> 19 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295091
#> 20 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594)  5121190
#> 21 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594)  5121193
#> 22 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295082
#> 23 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295093
#> 24 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594)  5121194
#> 25 [0000-0001-5483-4594](https://orcid.org/0000-0001-5483-4594) 46295092
#> 26 [0000-0001-9134-2871](https://orcid.org/0000-0001-9134-2871) 46348381
#> 27 [0000-0001-9134-2871](https://orcid.org/0000-0001-9134-2871) 46348378
#> 28 [0000-0001-9134-2871](https://orcid.org/0000-0001-9134-2871) 46348382
#> 29 [0000-0001-9134-2871](https://orcid.org/0000-0001-9134-2871) 46348383
#> 30 [0000-0001-9134-2871](https://orcid.org/0000-0001-9134-2871) 46348380
#> 31 [0000-0001-9134-2871](https://orcid.org/0000-0001-9134-2871) 46348377
#> 32 [0000-0001-9134-2871](https://orcid.org/0000-0001-9134-2871) 46348379
#> 33 [0000-0002-7436-8575](https://orcid.org/0000-0002-7436-8575)       NA
#> 34 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 46034653
#> 35 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 42201681
#> 36 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 42201680
#> 37 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 42186453
#> 38 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 42186456
#> 39 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 42186447
#> 40 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 26523259
#> 41 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 26523271
#> 42 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 26518384
#> 43 [0000-0003-0647-7726](https://orcid.org/0000-0003-0647-7726) 26523346
#>                type visibility                               path display-index
#> 1   JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295084             0
#> 2   JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295081             0
#> 3   JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295078             0
#> 4   JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295079             0
#> 5   JOURNAL_ARTICLE     PUBLIC  /0000-0001-5483-4594/work/5121191             0
#> 6   JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295089             0
#> 7   JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295080             0
#> 8  CONFERENCE_PAPER     PUBLIC /0000-0001-5483-4594/work/46295077             0
#> 9   JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295085             0
#> 10  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295087             0
#> 11  JOURNAL_ARTICLE     PUBLIC  /0000-0001-5483-4594/work/5121188             0
#> 12  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295086             0
#> 13  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295083             0
#> 14  JOURNAL_ARTICLE     PUBLIC  /0000-0001-5483-4594/work/5121192             0
#> 15  JOURNAL_ARTICLE     PUBLIC  /0000-0001-5483-4594/work/5121195             0
#> 16  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295090             0
#> 17  JOURNAL_ARTICLE     PUBLIC  /0000-0001-5483-4594/work/5121189             0
#> 18  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295088             0
#> 19  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295091             0
#> 20  JOURNAL_ARTICLE     PUBLIC  /0000-0001-5483-4594/work/5121190             0
#> 21  JOURNAL_ARTICLE     PUBLIC  /0000-0001-5483-4594/work/5121193             0
#> 22  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295082             0
#> 23  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295093             0
#> 24  JOURNAL_ARTICLE     PUBLIC  /0000-0001-5483-4594/work/5121194             0
#> 25  JOURNAL_ARTICLE     PUBLIC /0000-0001-5483-4594/work/46295092             0
#> 26  JOURNAL_ARTICLE     PUBLIC /0000-0001-9134-2871/work/46348381             0
#> 27  JOURNAL_ARTICLE     PUBLIC /0000-0001-9134-2871/work/46348378             0
#> 28  JOURNAL_ARTICLE     PUBLIC /0000-0001-9134-2871/work/46348382             0
#> 29  JOURNAL_ARTICLE     PUBLIC /0000-0001-9134-2871/work/46348383             0
#> 30  JOURNAL_ARTICLE     PUBLIC /0000-0001-9134-2871/work/46348380             0
#> 31 CONFERENCE_PAPER     PUBLIC /0000-0001-9134-2871/work/46348377             0
#> 32  JOURNAL_ARTICLE     PUBLIC /0000-0001-9134-2871/work/46348379             0
#> 33             <NA>       <NA>                               <NA>          <NA>
#> 34            OTHER     PUBLIC /0000-0003-0647-7726/work/46034653             0
#> 35            OTHER     PUBLIC /0000-0003-0647-7726/work/42201681             0
#> 36            OTHER     PUBLIC /0000-0003-0647-7726/work/42201680             0
#> 37            OTHER     PUBLIC /0000-0003-0647-7726/work/42186453             0
#> 38            OTHER     PUBLIC /0000-0003-0647-7726/work/42186456             0
#> 39            OTHER     PUBLIC /0000-0003-0647-7726/work/42186447             0
#> 40            OTHER     PUBLIC /0000-0003-0647-7726/work/26523259             0
#> 41            OTHER     PUBLIC /0000-0003-0647-7726/work/26523271             0
#> 42  JOURNAL_ARTICLE     PUBLIC /0000-0003-0647-7726/work/26518384             1
#> 43            OTHER     PUBLIC /0000-0003-0647-7726/work/26523346             0
#>    created-date.value last-modified-date.value source.source-orcid
#> 1        1.530808e+12             1.536155e+12                  NA
#> 2        1.530808e+12             1.536155e+12                  NA
#> 3        1.530808e+12             1.536155e+12                  NA
#> 4        1.530808e+12             1.536155e+12                  NA
#> 5        1.362697e+12             1.536155e+12                  NA
#> 6        1.530808e+12             1.536155e+12                  NA
#> 7        1.530808e+12             1.536155e+12                  NA
#> 8        1.530808e+12             1.536155e+12                  NA
#> 9        1.530808e+12             1.536155e+12                  NA
#> 10       1.530808e+12             1.536155e+12                  NA
#> 11       1.362697e+12             1.536155e+12                  NA
#> 12       1.530808e+12             1.536155e+12                  NA
#> 13       1.530808e+12             1.536155e+12                  NA
#> 14       1.362697e+12             1.536155e+12                  NA
#> 15       1.362697e+12             1.536155e+12                  NA
#> 16       1.530808e+12             1.536155e+12                  NA
#> 17       1.362697e+12             1.536155e+12                  NA
#> 18       1.530808e+12             1.536155e+12                  NA
#> 19       1.530808e+12             1.536155e+12                  NA
#> 20       1.362697e+12             1.536155e+12                  NA
#> 21       1.362697e+12             1.536155e+12                  NA
#> 22       1.530808e+12             1.536155e+12                  NA
#> 23       1.530808e+12             1.536155e+12                  NA
#> 24       1.362697e+12             1.536155e+12                  NA
#> 25       1.530808e+12             1.536155e+12                  NA
#> 26       1.530972e+12             1.530972e+12                  NA
#> 27       1.530972e+12             1.530972e+12                  NA
#> 28       1.530972e+12             1.530972e+12                  NA
#> 29       1.530972e+12             1.530972e+12                  NA
#> 30       1.530972e+12             1.530972e+12                  NA
#> 31       1.530972e+12             1.530972e+12                  NA
#> 32       1.530972e+12             1.530972e+12                  NA
#> 33                 NA                       NA                  NA
#> 34       1.530046e+12             1.530046e+12                  NA
#> 35       1.519912e+12             1.520426e+12                  NA
#> 36       1.519912e+12             1.520426e+12                  NA
#> 37       1.519891e+12             1.520426e+12                  NA
#> 38       1.519891e+12             1.520426e+12                  NA
#> 39       1.519891e+12             1.520426e+12                  NA
#> 40       1.473167e+12             1.520426e+12                  NA
#> 41       1.473167e+12             1.520426e+12                  NA
#> 42       1.473156e+12             1.520426e+12                  NA
#> 43       1.473167e+12             1.520426e+12                  NA
#>                      source.source-client-id.uri source.source-client-id.path
#> 1   https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 2   https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 3   https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 4   https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 5                                           <NA>                         <NA>
#> 6   https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 7   https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 8   https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 9   https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 10  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 11                                          <NA>                         <NA>
#> 12  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 13  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 14                                          <NA>                         <NA>
#> 15                                          <NA>                         <NA>
#> 16  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 17                                          <NA>                         <NA>
#> 18  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 19  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 20                                          <NA>                         <NA>
#> 21                                          <NA>                         <NA>
#> 22  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 23  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 24                                          <NA>                         <NA>
#> 25  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 26  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 27  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 28  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 29  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 30  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 31  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 32  https://orcid.org/client/0000-0002-5982-8983          0000-0002-5982-8983
#> 33                                          <NA>                         <NA>
#> 34 https://orcid.org/client/APP-HCGH49OO61VH1II3         APP-HCGH49OO61VH1II3
#> 35  https://orcid.org/client/0000-0001-8099-6984          0000-0001-8099-6984
#> 36  https://orcid.org/client/0000-0001-8099-6984          0000-0001-8099-6984
#> 37  https://orcid.org/client/0000-0001-8099-6984          0000-0001-8099-6984
#> 38  https://orcid.org/client/0000-0001-8099-6984          0000-0001-8099-6984
#> 39  https://orcid.org/client/0000-0001-8099-6984          0000-0001-8099-6984
#> 40  https://orcid.org/client/0000-0001-8099-6984          0000-0001-8099-6984
#> 41  https://orcid.org/client/0000-0001-8099-6984          0000-0001-8099-6984
#> 42                                          <NA>                         <NA>
#> 43  https://orcid.org/client/0000-0001-8099-6984          0000-0001-8099-6984
#>    source.source-client-id.host                source.source-name.value
#> 1                     orcid.org                       Scopus - Elsevier
#> 2                     orcid.org                       Scopus - Elsevier
#> 3                     orcid.org                       Scopus - Elsevier
#> 4                     orcid.org                       Scopus - Elsevier
#> 5                          <NA>                       Andreas Matzinger
#> 6                     orcid.org                       Scopus - Elsevier
#> 7                     orcid.org                       Scopus - Elsevier
#> 8                     orcid.org                       Scopus - Elsevier
#> 9                     orcid.org                       Scopus - Elsevier
#> 10                    orcid.org                       Scopus - Elsevier
#> 11                         <NA>                       Andreas Matzinger
#> 12                    orcid.org                       Scopus - Elsevier
#> 13                    orcid.org                       Scopus - Elsevier
#> 14                         <NA>                       Andreas Matzinger
#> 15                         <NA>                       Andreas Matzinger
#> 16                    orcid.org                       Scopus - Elsevier
#> 17                         <NA>                       Andreas Matzinger
#> 18                    orcid.org                       Scopus - Elsevier
#> 19                    orcid.org                       Scopus - Elsevier
#> 20                         <NA>                       Andreas Matzinger
#> 21                         <NA>                       Andreas Matzinger
#> 22                    orcid.org                       Scopus - Elsevier
#> 23                    orcid.org                       Scopus - Elsevier
#> 24                         <NA>                       Andreas Matzinger
#> 25                    orcid.org                       Scopus - Elsevier
#> 26                    orcid.org                       Scopus - Elsevier
#> 27                    orcid.org                       Scopus - Elsevier
#> 28                    orcid.org                       Scopus - Elsevier
#> 29                    orcid.org                       Scopus - Elsevier
#> 30                    orcid.org                       Scopus - Elsevier
#> 31                    orcid.org                       Scopus - Elsevier
#> 32                    orcid.org                       Scopus - Elsevier
#> 33                         <NA>                                    <NA>
#> 34                    orcid.org BASE - Bielefeld Academic Search Engine
#> 35                    orcid.org                                DataCite
#> 36                    orcid.org                                DataCite
#> 37                    orcid.org                                DataCite
#> 38                    orcid.org                                DataCite
#> 39                    orcid.org                                DataCite
#> 40                    orcid.org                                DataCite
#> 41                    orcid.org                                DataCite
#> 42                         <NA>                         Michael Rustler
#> 43                    orcid.org                                DataCite
#>    title.subtitle title.translated-title
#> 1              NA                     NA
#> 2              NA                     NA
#> 3              NA                     NA
#> 4              NA                     NA
#> 5              NA                     NA
#> 6              NA                     NA
#> 7              NA                     NA
#> 8              NA                     NA
#> 9              NA                     NA
#> 10             NA                     NA
#> 11             NA                     NA
#> 12             NA                     NA
#> 13             NA                     NA
#> 14             NA                     NA
#> 15             NA                     NA
#> 16             NA                     NA
#> 17             NA                     NA
#> 18             NA                     NA
#> 19             NA                     NA
#> 20             NA                     NA
#> 21             NA                     NA
#> 22             NA                     NA
#> 23             NA                     NA
#> 24             NA                     NA
#> 25             NA                     NA
#> 26             NA                     NA
#> 27             NA                     NA
#> 28             NA                     NA
#> 29             NA                     NA
#> 30             NA                     NA
#> 31             NA                     NA
#> 32             NA                     NA
#> 33             NA                     NA
#> 34             NA                     NA
#> 35             NA                     NA
#> 36             NA                     NA
#> 37             NA                     NA
#> 38             NA                     NA
#> 39             NA                     NA
#> 40             NA                     NA
#> 41             NA                     NA
#> 42             NA                     NA
#> 43             NA                     NA
#>                                                                                                                                                                                                         title.title.value
#> 1                                                                                                Impacts of combined sewer overflows on a large urban river – Understanding the effect of different management strategies
#> 2                                                                                                                                                                  Best data handling practices in water-related research
#> 3                                                                                                     Concurrent nitrate and atrazine retention in bioreactors of straw and bark mulch at short hydraulic residence times
#> 4                                                                                                    The influence of local calibration on the quality of UV-VIS spectrometer measurements in urban stormwater monitoring
#> 5                                                                                                                                                                        Hypolimnetic Oxygen Depletion in Eutrophic Lakes
#> 6                                                                                                                                                                        Hypolimnetic oxygen depletion in eutrophic lakes
#> 7  Potential for reducing nutrient input in surface waters by using nutrient filters,Potentiale für den einsatz von nährstoff-filtersystemen in Deutschland zur verringerung der nährstoffeinträge in oberflächengewässer
#> 8                                                                                                                         Validation and sensitivity of a coupled model tool for CSO impact assessment in Berlin, Germany
#> 9                                                                                                                              Development of a GIS method to localize critical source areas of diffuse nitrate pollution
#> 10                                                                                                                            Evaluation of effectiveness of combined sewer overflow control measures by operational data
#> 11                                                                                                                                   Contribution of combined sewer overflows to trace contaminant loads in urban streams
#> 12                                                                                                                                   Contribution of combined sewer overflows to trace contaminant loads in urban streams
#> 13                                                                                                                         Hypolimnetic oxygen consumption by sediment-based reduced substances in former eutrophic lakes
#> 14                                                                                                                         Hypolimnetic oxygen consumption by sediment-based reduced substances in former eutrophic lakes
#> 15                                                                                                                                          The potential of Lake Ohrid for long-term palaeoenvironmental reconstructions
#> 16                                                                                                                                          The potential of Lake Ohrid for long-term palaeoenvironmental reconstructions
#> 17                                                                                                                                              Effects of impoundment on nutrient availability and productivity in lakes
#> 18                                                                                                                                              Effects of impoundment on nutrient availability and productivity in lakes
#> 19                                                                                                        Eutrophication of ancient Lake Ohrid: Global warming amplifies detrimental effects of increased nutrient inputs
#> 20                                                                                                        Eutrophication of ancient Lake Ohrid: Global warming amplifies detrimental effects of increased nutrient inputs
#> 21                                                                                                                                    Sensitivity of Ancient Lake Ohrid to Local Anthropogenic Impacts and Global Warming
#> 22                                                                                                                                    Sensitivity of ancient Lake Ohrid to local anthropogenic impacts and global warming
#> 23                                                                                                                                                       Is Lake Prespa jeopardizing the ecosystem of ancient Lake Ohrid?
#> 24                                                                                                                              Similarities and differences in the annual temperature cycles of East German mining lakes
#> 25                                                                                                                              Similarities and differences in the annual temperature cycles of East German mining lakes
#> 26                                                                                                                                  The relevance of sewer deterioration modelling to support asset management strategies
#> 27                                                                                               Impacts of combined sewer overflows on a large urban river – Understanding the effect of different management strategies
#> 28                                                                                Influence of local calibration on the quality of online wet weather discharge monitoring: Feedback from five international case studies
#> 29                                                                                                                                                                 Best data handling practices in water-related research
#> 30                                                                                                   The influence of local calibration on the quality of UV-VIS spectrometer measurements in urban stormwater monitoring
#> 31                                                                                                                        Validation and sensitivity of a coupled model tool for CSO impact assessment in Berlin, Germany
#> 32                                                                                                                            Evaluation of effectiveness of combined sewer overflow control measures by operational data
#> 33                                                                                                                                                                                                                   <NA>
#> 34                                                                                                                                                                                               aquanes.report (v.0.5.0)
#> 35                                                                                                                                                                                               Aquanes.Report (V.0.4.0)
#> 36                                                                                                                Quantitative Microbiological Risk Assessment For Different Wastewater Reuse Options In Old Ford (V.1.0)
#> 37                                                                                                                                                                                                     Kwb.Qmra (V.0.1.1)
#> 38                                                                                                                                                                                                  Kwb.Hantush (V.0.2.1)
#> 39                                                                                                                                                                                                     Kwb.Wtaq (V.0.2.1)
#> 40                                                                                                                                                                                       Wrap your model in an R package!
#> 41                                                                                                                                                                                                       WRAP YOUR MODEL!
#> 42                                                                                                                                                                 Best data handling practices in water-related research
#> 43                                                                                                                                                         APPLICATION OF A DATA-DRIVEN APPROACH FOR WELL FIELD MODELLING
#>                                                                                                                                                                                                                   external-ids.external-id
#> 1                                                                                                                                                           doi, eid, 10.1016/j.watres.2016.08.017, 2-s2.0-84986616713, NA, NA, SELF, SELF
#> 2                                                                                                                                                                   doi, eid, 10.2166/wpt.2013.039, 2-s2.0-84890261205, NA, NA, SELF, SELF
#> 3                                                                                                                                                          doi, eid, 10.1016/j.ecoleng.2013.02.010, 2-s2.0-84875772175, NA, NA, SELF, SELF
#> 4                                                                                                                                                                   doi, eid, 10.2166/wpt.2013.042, 2-s2.0-84890261164, NA, NA, SELF, SELF
#> 5                                                                                                                                                                                                           doi, 10.1021/es301422r, SELF, 
#> 6                                                                                                                                                                      doi, eid, 10.1021/es301422r, 2-s2.0-84866433505, NA, NA, SELF, SELF
#> 7                                                                                                                                                                  doi, eid, 10.5675/HyWa_2012,1_1, 2-s2.0-84857542546, NA, NA, SELF, SELF
#> 8                                                                                                                                                                                                        eid, 2-s2.0-84894194693, NA, SELF
#> 9                                                                                                                                                                   doi, eid, 10.2166/wst.2011.672, 2-s2.0-80052089733, NA, NA, SELF, SELF
#> 10                                                                                                                                                                  doi, eid, 10.2166/wst.2011.058, 2-s2.0-79551713372, NA, NA, SELF, SELF
#> 11                                                                                                                                                                                               doi, 10.1016/j.watres.2010.06.011, SELF, 
#> 12                                                                                                                                                          doi, eid, 10.1016/j.watres.2010.06.011, 2-s2.0-77955267216, NA, NA, SELF, SELF
#> 13                                                                                                                                                             doi, eid, 10.4319/lo.2010.55.5.2073, 2-s2.0-77958043923, NA, NA, SELF, SELF
#> 14                                                                                                                                                                                                  doi, 10.4319/lo.2010.55.5.2073, SELF, 
#> 15                                                                                                                                                                                               doi, 10.1016/j.palaeo.2007.10.015, SELF, 
#> 16                                                                                                                                                          doi, eid, 10.1016/j.palaeo.2007.10.015, 2-s2.0-40149110826, NA, NA, SELF, SELF
#> 17                                                                                                                                                                                                  doi, 10.4319/lo.2007.52.6.2629, SELF, 
#> 18                                                                                                                                                             doi, eid, 10.4319/lo.2007.52.6.2629, 2-s2.0-36749100518, NA, NA, SELF, SELF
#> 19                                                                                                                                                             doi, eid, 10.4319/lo.2007.52.1.0338, 2-s2.0-34250897864, NA, NA, SELF, SELF
#> 20                                                                                                                                                                                                  doi, 10.4319/lo.2007.52.1.0338, SELF, 
#> 21                                                                                                                                                                              doi, 10.3394/0380-1330(2006)32[158:SOALOT]2.0.CO;2, SELF, 
#> 22                                                                                                                                         doi, eid, 10.3394/0380-1330(2006)32[158:SOALOT]2.0.CO;2, 2-s2.0-33645746374, NA, NA, SELF, SELF
#> 23                                                                                                                                                             doi, eid, 10.1007/s10750-005-6427-9, 2-s2.0-28444461482, NA, NA, SELF, SELF
#> 24                                                                                                                                                                                              doi, 10.1016/S0075-9511(00)80058-4, SELF, 
#> 25                                                                                                                                                          doi, eid, 10.1016/S0075-9511(00)80058-4, 2-s2.0-0033793580, NA, NA, SELF, SELF
#> 26                                                                                                                                                         doi, eid, 10.1080/1573062X.2017.1325497, 2-s2.0-85019615780, NA, NA, SELF, SELF
#> 27                                                                                                                                                          doi, eid, 10.1016/j.watres.2016.08.017, 2-s2.0-84986616713, NA, NA, SELF, SELF
#> 28                                                                                                                                                                  doi, eid, 10.2166/wst.2014.465, 2-s2.0-84923474533, NA, NA, SELF, SELF
#> 29                                                                                                                                                                  doi, eid, 10.2166/wpt.2013.039, 2-s2.0-84890261205, NA, NA, SELF, SELF
#> 30                                                                                                                                                                  doi, eid, 10.2166/wpt.2013.042, 2-s2.0-84890261164, NA, NA, SELF, SELF
#> 31                                                                                                                                                                                                       eid, 2-s2.0-84894194693, NA, SELF
#> 32                                                                                                                                                                  doi, eid, 10.2166/wst.2011.058, 2-s2.0-79551713372, NA, NA, SELF, SELF
#> 33                                                                                                                                                                                                                                    NULL
#> 34 source-work-id, doi, BASE:9e86b2ebd511f53078a854c3b5a5840fe523a63eb997a8a89fbd04c92f08b0f5, 10.5281/zenodo.1244103, SELF, SELF, https://www.base-search.net/Record/9e86b2ebd511f53078a854c3b5a5840fe523a63eb997a8a89fbd04c92f08b0f5, NA
#> 35                                                                                                                                                                                                    doi, 10.5281/zenodo.825030, NA, SELF
#> 36                                                                                                                                                                                                    doi, 10.5281/zenodo.159527, NA, SELF
#> 37                                                                                                                                                                                                    doi, 10.5281/zenodo.154111, NA, SELF
#> 38                                                                                                                                                                                                     doi, 10.5281/zenodo.61613, NA, SELF
#> 39                                                                                                                                                                                                     doi, 10.5281/zenodo.61610, NA, SELF
#> 40                                                                                                                                                                                                     doi, 10.5281/ZENODO.58029, NA, SELF
#> 41                                                                                                                                                                                                doi, 10.13140/RG.2.1.2140.3683, NA, SELF
#> 42                                                                                                                                                                     doi, 10.2166/wpt.2013.039, SELF, http://doi.org/10.2166/wpt.2013.03
#> 43                                                                                                                                                                                                doi, 10.13140/RG.2.1.3151.5688, NA, SELF
#>    publication-date.month publication-date.day publication-date.media-type
#> 1                      NA                   NA                          NA
#> 2                      NA                   NA                          NA
#> 3                      NA                   NA                          NA
#> 4                      NA                   NA                          NA
#> 5                      NA                   NA                          NA
#> 6                      NA                   NA                          NA
#> 7                      NA                   NA                          NA
#> 8                      NA                   NA                          NA
#> 9                      NA                   NA                          NA
#> 10                     NA                   NA                          NA
#> 11                     NA                   NA                          NA
#> 12                     NA                   NA                          NA
#> 13                     NA                   NA                          NA
#> 14                     NA                   NA                          NA
#> 15                     NA                   NA                          NA
#> 16                     NA                   NA                          NA
#> 17                     NA                   NA                          NA
#> 18                     NA                   NA                          NA
#> 19                     NA                   NA                          NA
#> 20                     NA                   NA                          NA
#> 21                     NA                   NA                          NA
#> 22                     NA                   NA                          NA
#> 23                     NA                   NA                          NA
#> 24                     NA                   NA                          NA
#> 25                     NA                   NA                          NA
#> 26                     NA                   NA                          NA
#> 27                     NA                   NA                          NA
#> 28                     NA                   NA                          NA
#> 29                     NA                   NA                          NA
#> 30                     NA                   NA                          NA
#> 31                     NA                   NA                          NA
#> 32                     NA                   NA                          NA
#> 33                     NA                   NA                          NA
#> 34                     NA                   NA                          NA
#> 35                     NA                   NA                          NA
#> 36                     NA                   NA                          NA
#> 37                     NA                   NA                          NA
#> 38                     NA                   NA                          NA
#> 39                     NA                   NA                          NA
#> 40                     NA                   NA                          NA
#> 41                     NA                   NA                          NA
#> 42                     NA                   NA                          NA
#> 43                     NA                   NA                          NA
#>    publication-date.year.value               source.source-orcid.uri
#> 1                         2016                                  <NA>
#> 2                         2013                                  <NA>
#> 3                         2013                                  <NA>
#> 4                         2013                                  <NA>
#> 5                         2012 https://orcid.org/0000-0001-5483-4594
#> 6                         2012                                  <NA>
#> 7                         2012                                  <NA>
#> 8                         2012                                  <NA>
#> 9                         2011                                  <NA>
#> 10                        2011                                  <NA>
#> 11                        2010 https://orcid.org/0000-0001-5483-4594
#> 12                        2010                                  <NA>
#> 13                        2010                                  <NA>
#> 14                        2010 https://orcid.org/0000-0001-5483-4594
#> 15                        2008 https://orcid.org/0000-0001-5483-4594
#> 16                        2008                                  <NA>
#> 17                        2007 https://orcid.org/0000-0001-5483-4594
#> 18                        2007                                  <NA>
#> 19                        2007                                  <NA>
#> 20                        2007 https://orcid.org/0000-0001-5483-4594
#> 21                        2006 https://orcid.org/0000-0001-5483-4594
#> 22                        2006                                  <NA>
#> 23                        2006                                  <NA>
#> 24                        2000 https://orcid.org/0000-0001-5483-4594
#> 25                        2000                                  <NA>
#> 26                        2017                                  <NA>
#> 27                        2016                                  <NA>
#> 28                        2015                                  <NA>
#> 29                        2013                                  <NA>
#> 30                        2013                                  <NA>
#> 31                        2012                                  <NA>
#> 32                        2011                                  <NA>
#> 33                        <NA>                                  <NA>
#> 34                        2018                                  <NA>
#> 35                        2017                                  <NA>
#> 36                        2016                                  <NA>
#> 37                        2016                                  <NA>
#> 38                        2016                                  <NA>
#> 39                        2016                                  <NA>
#> 40                        2016                                  <NA>
#> 41                        2014                                  <NA>
#> 42                        2013 https://orcid.org/0000-0003-0647-7726
#> 43                        2012                                  <NA>
#>    source.source-orcid.path source.source-orcid.host
#> 1                      <NA>                     <NA>
#> 2                      <NA>                     <NA>
#> 3                      <NA>                     <NA>
#> 4                      <NA>                     <NA>
#> 5       0000-0001-5483-4594                orcid.org
#> 6                      <NA>                     <NA>
#> 7                      <NA>                     <NA>
#> 8                      <NA>                     <NA>
#> 9                      <NA>                     <NA>
#> 10                     <NA>                     <NA>
#> 11      0000-0001-5483-4594                orcid.org
#> 12                     <NA>                     <NA>
#> 13                     <NA>                     <NA>
#> 14      0000-0001-5483-4594                orcid.org
#> 15      0000-0001-5483-4594                orcid.org
#> 16                     <NA>                     <NA>
#> 17      0000-0001-5483-4594                orcid.org
#> 18                     <NA>                     <NA>
#> 19                     <NA>                     <NA>
#> 20      0000-0001-5483-4594                orcid.org
#> 21      0000-0001-5483-4594                orcid.org
#> 22                     <NA>                     <NA>
#> 23                     <NA>                     <NA>
#> 24      0000-0001-5483-4594                orcid.org
#> 25                     <NA>                     <NA>
#> 26                     <NA>                     <NA>
#> 27                     <NA>                     <NA>
#> 28                     <NA>                     <NA>
#> 29                     <NA>                     <NA>
#> 30                     <NA>                     <NA>
#> 31                     <NA>                     <NA>
#> 32                     <NA>                     <NA>
#> 33                     <NA>                     <NA>
#> 34                     <NA>                     <NA>
#> 35                     <NA>                     <NA>
#> 36                     <NA>                     <NA>
#> 37                     <NA>                     <NA>
#> 38                     <NA>                     <NA>
#> 39                     <NA>                     <NA>
#> 40                     <NA>                     <NA>
#> 41                     <NA>                     <NA>
#> 42      0000-0003-0647-7726                orcid.org
#> 43                     <NA>                     <NA>
#>    publication-date.month.value publication-date.day.value
#> 1                          <NA>                       <NA>
#> 2                          <NA>                       <NA>
#> 3                          <NA>                       <NA>
#> 4                          <NA>                       <NA>
#> 5                            08                         24
#> 6                          <NA>                       <NA>
#> 7                          <NA>                       <NA>
#> 8                          <NA>                       <NA>
#> 9                          <NA>                       <NA>
#> 10                         <NA>                       <NA>
#> 11                           08                       <NA>
#> 12                         <NA>                       <NA>
#> 13                         <NA>                       <NA>
#> 14                         <NA>                       <NA>
#> 15                           03                       <NA>
#> 16                         <NA>                       <NA>
#> 17                         <NA>                       <NA>
#> 18                         <NA>                       <NA>
#> 19                         <NA>                       <NA>
#> 20                         <NA>                       <NA>
#> 21                           03                       <NA>
#> 22                         <NA>                       <NA>
#> 23                         <NA>                       <NA>
#> 24                           09                       <NA>
#> 25                         <NA>                       <NA>
#> 26                         <NA>                       <NA>
#> 27                         <NA>                       <NA>
#> 28                         <NA>                       <NA>
#> 29                         <NA>                       <NA>
#> 30                         <NA>                       <NA>
#> 31                         <NA>                       <NA>
#> 32                         <NA>                       <NA>
#> 33                         <NA>                       <NA>
#> 34                         <NA>                       <NA>
#> 35                           07                         10
#> 36                           10                         07
#> 37                           09                         15
#> 38                           09                         06
#> 39                           09                         06
#> 40                         <NA>                       <NA>
#> 41                         <NA>                       <NA>
#> 42                         <NA>                       <NA>
#> 43                         <NA>                       <NA>
#>    source.source-client-id
#> 1                       NA
#> 2                       NA
#> 3                       NA
#> 4                       NA
#> 5                       NA
#> 6                       NA
#> 7                       NA
#> 8                       NA
#> 9                       NA
#> 10                      NA
#> 11                      NA
#> 12                      NA
#> 13                      NA
#> 14                      NA
#> 15                      NA
#> 16                      NA
#> 17                      NA
#> 18                      NA
#> 19                      NA
#> 20                      NA
#> 21                      NA
#> 22                      NA
#> 23                      NA
#> 24                      NA
#> 25                      NA
#> 26                      NA
#> 27                      NA
#> 28                      NA
#> 29                      NA
#> 30                      NA
#> 31                      NA
#> 32                      NA
#> 33                      NA
#> 34                      NA
#> 35                      NA
#> 36                      NA
#> 37                      NA
#> 38                      NA
#> 39                      NA
#> 40                      NA
#> 41                      NA
#> 42                      NA
#> 43                      NA